script/requirements.txt
numpy>=1.20.0
pillow>=9.0.0
earthengine-api>=1.7.0
requests>=2.28.0
scipy>=1.7.0
tqdm>=4.65.0
